requirements.txt
psycopg2-binary==2.9.10
pydantic==2.11.7
boto3>=1.35.0
pgvector==0.3.6
